+++ CHANGES.md
@@ -5,6 +5,7 @@ Features ✨:
  -
 
 Improvements 🙌:
+ - Share image and other media from e2e rooms (#677)
  - Add support for `/plain` command (#12)
  - Detect spaces in password if user fail to login (#1038)
  - FTUE: do not display a different color when encrypting message when not in developer mode.
@@ -12,6 +13,8 @@ Improvements 🙌:
 
 Bugfix 🐛:
  - Fix crash on attachment preview screen (#1088)
+ - "Share" option is not appearing in encrypted rooms for images (#1031)
+ - Set "image/jpeg" as MIME type of images instead of "image/jpg" (#1075)
 
 Translations 🗣:
  -